cleaned up code for pull request
